Use resampled coordinates in ACTION_UP
Currently, resampled coordinates only affect ACTION_MOVE events.
However, this can cause the following scenario:
1) The past two motion events were batched and resulted in a resampling.
So a resampled coordinate is reported.
2) The next event is ACTION_UP, which reports "real" coordinates

The situation here could lead to an unwanted "jump back" in coordinates.
If all motion events are occuring with linear speed, then resampling
would extrapolate the coordinates into the future. But then ACTION_UP
would use an old un-extrapolated coordinate, and the motion would become
non-monotonic.

This could be especially significant in tests, where the events are
being injected with a very predictable schedule, but are not aligned
with VSYNC, and therefore can be batched in unpredictable manner. That
means that some instances of the test could potentially cause the motion
back to occur, while others would not. If the test relies on this
behaviour, this would cause flakes in the tests.

Bug: 119214052
Test: atest \
  InputPublisherAndConsumerTest#PublishMotionEvent_ActionUpGetsResampled
Test: atest \
  InputPublisherAndConsumerTest#PublishMotionEvent_ActionPointerUpGetsResampled

Flag: com.android.input.flags.fix_action_up_resampling

Change-Id: I29e662b1a506f74c6e305c445d8d65061576e0ba