SF: Ignore backpressure for lockstep followers
If the pacesetter and follower have equal or roughly equal refresh rates
(e.g. 60 Hz and 59.94 Hz), PDARR enabled the follower to skip presenting
to avoid backpressure, but did not force it to be composited on the GPU.
For DPU-composited buffers mirrored on multiple displays, PDARR does not
yet do the bookkeeping to merge release fences over consecutive frames,
so on frames skipped by the follower, buffers were released prematurely.

Fix this by always presenting the follower in that case, which restores
the previous behavior of not recovering from backpressure, but allowing
DPU composition. In other words, tie the following PDARR flags:

    follower_display_backpressure
    force_slower_follower_gpu_composition

Also, clean up unnecessary use of mStateLock and hasActiveMode.

Fixes: 461663004
Flag: EXEMPT BUGFIX
Test: No green artifacts while mirroring.
Test: New SchedulerTest
Change-Id: I6c806ec77219a6c18bcf9c918c4b29f8792acfc9