Move ScopedFlagOverride.h to include/input
SCOPED_FLAG_OVERRIDE is a macro which is useful not only for inputflinger tests but also other input tests. This includes InputConsumer tests which live in libs/input. We move this into a more general shared directory such that we can use this macro more generally. It makes sense to put this in input/ as the code is implemented specifically for input_flags.

Test: atest --host libinput_tests inputflinger_tests
Flag: EXEMPT PURE_REFACTOR
Bug: 466004548
Change-Id: Ica2f3aec0a6358d55769c7664bed9ab96895e455