Fix getExtension for BinderRpc
Attempts to call getExtension fail because the Parcel that is being created for the EXTENSION_TRANSACTION is not marked for rpc. Fix that and add a test.

Exempt bugfix rationale: the fix is thoroughly tested and the existence of the bug proves there's no existing code that depends on it (it is impossible to call getExtensions on an RpcBinder as it currently stands).

Bug: 462787437
Test: atest binderRpcTest
Flag: EXEMPT BUGFIX
Change-Id: I817c8bc49146d479bea54a1232eb17e60579c99a