Pass JavaVM to input stages and InputThread
Before this CL, JniEnv was passed throughout input code. However, it's
expected that only an attached thread should interact with the env.

To avoid the incorrect thread access, we obtain the JavaVm and pass that
around instead. It can then be used to attach the thread of interest to
Java.

Bug: 463458159
Test: ensured device boots
Flag: EXEMPT PURE_REFACTOR
Change-Id: I0abd9e04a720a7636aeb7cc2752d7a9200209119