BufferQueue: Prevent AB-BA deadlocks by making free callbacks w/o locks
This change refactors the buffer freeing mechanism across BufferQueue
consumers to prevent AB-BA deadlocks. Previously, `onBufferFreed`
callbacks were invoked while holding mutex locks, which could lead to
deadlocks if the callback itself attempted to acquire other locks alrea
held elsewhere.

The change also gives BufferItemConsumer methods an optional callback
that can be used in lieu of the listener. That way, the callback can
handle the changes assuming the lock is held, instead of
BufferItemConsumer trying to call a callback that takes a lock during a
call that has the lock.

Bug: 447613760
Flag: EXEMPT BUGFIX
Test: atest, new tests
Change-Id: I15341b112fd74429f53831af13fa67d6cc4c357f