Add AServiceManager_registerLazyServiceWithFlags to llndk
This is the same as the existing API AServiceManager_registerLazyService
but now **with flags**!

This allows lazy services in /vendor to be registered with these flags.
The same flags are available for AServiceManager_addService, so this is
adding the same functionality for lazy services.

Note: Flags aren't available for llndk APIs. Clients that use the new
APIs can be flagged. The existance of the API doesn't change behavior
without users of it.

Test: libbinder_ndk_unit_tests
Bug: 449206267
Flag: EXEMPT BUGFIX
Change-Id: Ib20e232bbde5cad2514a71de451fd667184645e4