Add call to syslog in fuzzer to create FD
We check for leaked FDs, so we need to set up known FDs before taking a
snapshot of the number of FDs.

syslog creates a SOCK_DGRAM when called from binder_parcel_utils.h so we
need to add a call to that before getting the initial number of FDs.

Test: binder_parcel_fuzzer
Bug: 468335028
Flag: EXEMPT BUGFIX
Change-Id: Iccf92f1351db56ee5230f5af1861182ee6055361