Filter out windows with the same token when hit testing spy windows
If a window and its clone are present, and both are SPY, InputDispatcher
may try to add both windows to the touch targets for this gesture.
This will cause a crash in InputDispatcher, due to overlap with incoming
pointers.

To fix this, we filter out spy windows with the same token.

Bug: 468885908
Test: TEST=inputflinger_tests; m $TEST && $ANDROID_HOST_OUT/nativetest64/$TEST/$TEST --gtest_filter="*TapClonedSpyWindow"
Flag: EXEMPT BUGFIX

Change-Id: I54a9776ad71e994282b3e05495ea116917835e36