Refactor bindgen deps for no_std support
Refactor both `libbinder_ndk_compat_bindgen` and `libbinder_ndk_bindgen`
so that they are supported in `libbinder_rs_nostd`. In the former case
this involves introducing a new feature, `"android_compat_symbols"`, to
replace the previous `"std"` and `trusty` guarded configurations. In the
latter case, this removes the `*_no_std` specific bindgen rules.

This lays the groundwork for expanding the number of APIs that are
provided through `libbinder_ndk_compat_bindgen` in upcoming changes.

Bug: 461767790
Test: atest rustBinderTest rustBinderTest_nostd binderRustNdkInteropTest
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ie7fee1554e10d9db5f5c39c583944b55ee7ae550