CompositionFilterFlag and exclusionMask for screenshots.
Introduce a new field called compositionFilterFlag in layer_state_t,
it's for categorizing layers that share the same type or behavior in
screen capture. It represents the properties of layers which help
SurfaceFlinger apply logic to specific layers during composition.

Also introduce exclusionMask in CaptureArgs. It's a bit mask for
excluding layers in screenshots. When a layer's composition filter AND
the exclusionMask is non-zero, the layer will be hidden from the
screenshots.

Test: LayerSnapshotTest and ScreenCaptureTest
Flag: com.android.window.flags.screen_capture_exclusion_flags
Bug: 469847524

Change-Id: I761a44481f4a858045b6f6d8a1c97ef430582957