Introduce TestEventQueue in InputMapper unit tests
This new queue acts similar to an InputListener stage. The
events emitted by "process" and by other functions will be added to the
queue, and the assertion can happen against the queue.

This simplifies the syntax in unit tests, and makes the tests more
similar to the other ones we have in the rest of the input code.

Bug: 283812079
Test: atest --host inputflinger_tests
Flag: EXEMPT TEST_ONLY
Change-Id: I4a25f6fc8933dc6d7282ba32d50f82157b9dcb2a