Add synchronization for transaction callbacks in BBQ dtor.
Linked bug points to memory corruption due to a data race.
BBQ::acquireNextBufferLocked(mSyncTransaction) is called with mMutex
held, however BBQ dtor might be running in another thread, calling the
callback which frees mSyncTransaction.  This patch attempts to prevent
this.  Consistent with other functions in BBQ, the callback is not
itself called under lock, only the pointers are switched out while the
lock is held.

Bug=428619795
Test: presubmit
Flag: EXEMPT BUGFIX

Change-Id: I2e4e2163e38e1084bee930a79fa6a0ed0fd99bd9